use int64_t for octave_hdf5_id (bug #47858)
* oct-hdf5-types.h (octave_hdf5_id): Use int64_t.

* ls-hdf5.cc (check_hdf5_id_value): New static function.
* ls-hdf5.h, ls-hdf5.cc (hdf5_read_next_data_internal): Rename from
hdf5_read_next_data.  Use HDF5 types directly.
(hdf5_read_next_data): Now a wrapper around
hdf5_read_next_data_internal.  Use octave HDF5 types in the interface.
(hdf5_h5g_iterate): New function.
(read_hdf5_data): Pass hdf5_read_next_data_internal instead of
hdf5_read_next_data to H5Giterate.

* ov-cell.cc (octave_cell::load_hdf5): Call hdf5_h5g_iterate wrapper
function instead of calling H5giterate directly.
* ov-class.cc (octave_class::load_hdf5): Likewise.
* ov-fcn-handle.cc (octave_fcn_handle::load_hdf5): Likewise.
* ov-struct.cc (octave_struct::load_hdf5): Likewise.

+++ libinterp/corefcn/oct-hdf5-types.h
@@ -32,7 +32,7 @@ extern bool check_hdf5_types (bool warn = true);
 
 // Available for C and C++.
 
-typedef int octave_hdf5_id;
+typedef int64_t octave_hdf5_id;
 typedef int octave_hdf5_err;
 
 #if defined (__cplusplus)
